+++ crates/protocol/src/platform.rs
@@ -16,33 +16,79 @@ use alloc::borrow::Cow;
 
 use wasefire_common::platform::Side;
 use wasefire_error::Error;
-use wasefire_wire::Wire;
+use wasefire_wire::{Wire, Yoke};
+
+#[derive(Debug)]
+pub enum DynInfo {
+    V2(Yoke<Info<'static>>),
+    V1(Yoke<_Info1<'static>>),
+    V0(Yoke<_Info0<'static>>),
+}
+
+#[cfg(feature = "host")]
+impl DynInfo {
+    pub fn serial(&self) -> &[u8] {
+        match self {
+            DynInfo::V2(x) => &x.get().serial,
+            DynInfo::V1(x) => &x.get().serial,
+            DynInfo::V0(x) => &x.get().serial,
+        }
+    }
+
+    pub fn applet_kind(&self) -> Option<AppletKind> {
+        match self {
+            DynInfo::V2(x) => Some(x.get().applet_kind),
+            DynInfo::V1(_) => None,
+            DynInfo::V0(_) => None,
+        }
+    }
+
+    pub fn running_side(&self) -> Option<Side> {
+        match self {
+            DynInfo::V2(x) => Some(x.get().running_side),
+            DynInfo::V1(x) => Some(x.get().running_side),
+            DynInfo::V0(_) => None,
+        }
+    }
+
+    pub fn running_version(&self) -> &[u8] {
+        match self {
+            DynInfo::V2(x) => &x.get().running_version,
+            DynInfo::V1(x) => &x.get().running_version,
+            DynInfo::V0(x) => &x.get().version,
+        }
+    }
+
+    pub fn opposite_version(&self) -> Option<Result<&[u8], Error>> {
+        let opposite = match self {
+            DynInfo::V2(x) => &x.get().opposite_version,
+            DynInfo::V1(x) => &x.get().opposite_version,
+            DynInfo::V0(_) => return None,
+        };
+        Some(match opposite {
+            Ok(x) => Ok(x),
+            Err(e) => Err(*e),
+        })
+    }
+}
 
 #[derive(Debug, Wire)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Info<'a> {
     pub serial: Cow<'a, [u8]>,
+    pub applet_kind: AppletKind,
     pub running_side: Side,
     pub running_version: Cow<'a, [u8]>,
     pub opposite_version: Result<Cow<'a, [u8]>, Error>,
 }
 
-#[cfg(feature = "host")]
-impl core::fmt::Display for Info<'_> {
-    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        use data_encoding::HEXLOWER as HEX;
-        writeln!(f, "  serial: {}", HEX.encode_display(&self.serial))?;
-        writeln!(f, "    side: {}", self.running_side)?;
-        writeln!(f, " version: {}", HEX.encode_display(&self.running_version))?;
-        writeln!(
-            f,
-            "opposite: {}",
-            match &self.opposite_version {
-                Ok(x) => HEX.encode(x),
-                Err(e) => alloc::format!("{e}"),
-            }
-        )
-    }
+#[derive(Debug, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct _Info1<'a> {
+    pub serial: Cow<'a, [u8]>,
+    pub running_side: Side,
+    pub running_version: Cow<'a, [u8]>,
+    pub opposite_version: Result<Cow<'a, [u8]>, Error>,
 }
 
 #[derive(Debug, Wire)]
@@ -50,4 +96,28 @@ impl core::fmt::Display for Info<'_> {
 pub struct _Info0<'a> {
     pub serial: Cow<'a, [u8]>,
     pub version: Cow<'a, [u8]>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Wire)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum AppletKind {
+    Wasm,
+    Pulley,
+    Native,
+}
+
+impl AppletKind {
+    pub fn name(&self) -> &'static str {
+        match self {
+            AppletKind::Wasm => "wasm",
+            AppletKind::Pulley => "pulley",
+            AppletKind::Native => "native",
+        }
+    }
+}
+
+impl core::fmt::Display for AppletKind {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        self.name().fmt(f)
+    }
 }